request data from mock ok

+++ alg/backtrace/backtrace_ans.go
@@ -147,7 +147,7 @@ func solve2(queens []int, n, row, columns, diagonals1, diagonals2 int) {
 		solutions = append(solutions, board)
 		return
 	}
-	availablePositions := ((1 << n) - 1) & (^(columns | diagonals1 | diagonals2))
+	availablePositions := ((1 << uint(n)) - 1) & (^(columns | diagonals1 | diagonals2))
 	for availablePositions != 0 {
 		position := availablePositions & (-availablePositions)
 		availablePositions = availablePositions & (availablePositions - 1)
